Add tests for reconstruction accuracy

+++ src/branch_sequences.c
@@ -562,8 +562,8 @@ int extend_lower_part_of_window(int starting_coord, int initial_min_coord, int g
 int get_blocks(int ** block_coordinates, int genome_size,int * snp_site_coords,int number_of_branch_snps, int window_size, int cutoff, char * original_sequence, int * snp_locations, int number_of_snps)
 {
 	// Set up the window counter with 1 value per base in the branch
-    int8_t * window_count;
-	window_count = (int8_t *) calloc((genome_size+1),sizeof(int8_t));
+    int * window_count;
+    window_count = (int *) calloc((genome_size+1),sizeof(int));
 	
 	// Integer array with location of gaps
     int8_t * gaps_in_original_genome_space;
